+++ src/index.js
@@ -1,5 +1,3 @@
-const semver = require('semver')
-const tailwindPkg = require('tailwindcss/package.json')
 const plugin = require('tailwindcss/plugin')
 
 const baseStyles = {
@@ -10,17 +8,6 @@ const baseStyles = {
 
 const lineClamp = plugin(
   function ({ matchUtilities, addUtilities, theme, variants }) {
-    if (semver.gte(tailwindPkg.version, '3.3.0')) {
-      console.log(
-        `\u001b[31m${[
-          'As of Tailwind CSS v3.3, the `@tailwindcss/line-clamp` plugin is now included by default.',
-          'Remove it from the `plugins` array in your configuration to eliminate this warning.',
-        ].join('\n')}\u001b[0m`
-      )
-
-      return
-    }
-
     const values = theme('lineClamp')
 
     matchUtilities(